solid/SRP/violation/requirements.txt
# Pinned requirements for reproducibility
torch==2.4.1
pyyaml==6.0.2
pandas==2.2.3
scikit-learn==1.5.2

